Move azure utils to utils dir.
And rename module similar to ec2 and gce utils format.
