+++ mash/services/replication/azure_job.py
@@ -16,7 +16,7 @@
 # along with mash.  If not, see <http://www.gnu.org/licenses/>
 #
 
-from mash.services.azure_utils import (
+from mash.utils.azure import (
     copy_blob_to_classic_storage,
     delete_image,
     delete_page_blob
