+++ mash/utils/azure.py
